Curve control
Now you can control the resolution and smoothing of the average curve. This is especially important when you have few tracks to compare, so naturally averaging alone won't lead to a smooth curve. On the other hand, if you have compared many tracks, like 10 or more, you don't want too smoothen the result too much because those spikes that survive that much averaging are clearly significant.